requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djoser==1.1.5
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.2
openapi-codec==1.3.2
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
